requirements.txt
Flask==1.1.1
pandas==0.25.0
google-cloud-firestore==1.4.0